# packages/api-client/openapi.json
{
  "components": {
    "responses": {
      "DEFAULT_ERROR": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Default error response"
      },
      "UNPROCESSABLE_ENTITY": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Unprocessable Entity"
      }
    },
    "schemas": {
...
      "AnalysisData": {
        "additionalProperties": false,
        "properties": {
          "analysis_mode": {
            "type": "string"
          },
          "completed_at": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "evidence_summary": {
            "$ref": "#/components/schemas/AnalysisEvidenceSummary"
          },
          "evidence_url": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "ocr_confirmation_id": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "ocr_result_id": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "ocr_review": {
            "$ref": "#/components/schemas/AnalysisOCRReview"
          },
          "progress": {
            "$ref": "#/components/schemas/AnalysisProgress"
          },
          "risk": {
            "$ref": "#/components/schemas/AnalysisRisk"
          },
          "status": {
            "type": "string"
          },
          "transaction_id": {
            "format": "uuid",
            "type": "string"
          },
          "verification": {
            "anyOf": [
              {
                "nullable": true,
                "type": "object"
              },
              {
                "$ref": "#/components/schemas/VerificationSummary"
              }
            ]
          },
          "versions": {
            "$ref": "#/components/schemas/AnalysisVersions"
          }
        },
        "required": [
          "analysis_mode",
          "completed_at",
          "created_at",
          "evidence_summary",
          "evidence_url",
          "id",
          "ocr_confirmation_id",
          "ocr_result_id",
          "ocr_review",
          "progress",
          "risk",
          "status",
          "transaction_id",
          "verification",
          "versions"
        ],
        "type": "object"
      },
      "AnalysisEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AnalysisData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "AnalysisStartEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AnalysisStartData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "AnalysisStartRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "default": "combined",
            "enum": [
              "combined",
              "screenshot_only"
            ],
            "type": "string"
          },
          "ocr_result_id": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Error code",
            "type": "integer"
          },
          "errors": {
            "additionalProperties": {},
            "description": "Errors",
            "type": "object"
          },
          "message": {
            "description": "Error message",
            "type": "string"
          },
          "status": {
            "description": "Error name",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Meta": {
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "type": "object"
      },
...
      "OCRConfirmationEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OCRConfirmationData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "OCRConfirmationRequest": {
        "additionalProperties": false,
        "properties": {
          "correction_reasons": {
            "additionalProperties": {
              "maxLength": 300,
              "minLength": 5,
              "type": "string"
            },
            "type": "object"
          },
          "fields": {
            "additionalProperties": {},
            "type": "object"
          },
          "ocr_result_id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "fields",
          "ocr_result_id"
        ],
        "type": "object"
      },
...
      "OCRReviewEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OCRReviewData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "TransactionHistoryEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionHistoryData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "TransactionUploadEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionUploadData"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MoMo-FDVS API",
    "version": "v1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/analyses/{analysis_run_id}": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Authentication is required."
          },
          "404": {
            "description": "Analysis is not visible to the caller."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Return the persisted policy result and bounded progress projection.",
        "tags": [
          "analyses-v1"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "analysis_run_id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ]
    },
...
    "/api/v1/transactions": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "DRAFT",
                "UPLOADED",
                "OCR_PENDING",
                "OCR_REVIEW",
                "READY",
                "ANALYSIS_QUEUED",
                "ANALYSING",
                "COMPLETED",
                "PARTIAL",
                "FAILED"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "verification",
            "required": false,
            "schema": {
              "enum": [
                "VERIFIED",
                "UNVERIFIED",
                "MISMATCH",
                "NOT_ATTEMPTED"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "band",
            "required": false,
            "schema": {
              "enum": [
                "low_risk",
                "medium_risk",
                "high_risk",
                "inconclusive"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionHistoryEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "List only the caller's transactions using persisted latest-run summaries.",
        "tags": [
          "transactions-v1"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "client_captured_at": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "client_metadata": {
                    "description": "A size-limited JSON object of safe scalar values.",
                    "type": "string"
                  },
                  "receipt": {
                    "format": "binary",
                    "type": "string"
                  },
                  "source": {
                    "enum": [
                      "CAMERA",
                      "GALLERY"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "receipt",
                  "source"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionUploadEnvelope"
                }
              }
            },
            "description": "A completed idempotent upload was replayed."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionUploadEnvelope"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Invalid image metadata or idempotency key."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The USER role is required."
          },
          "409": {
            "description": "Idempotency conflict."
          },
          "413": {
            "description": "Encoded bytes or decoded pixels exceed a limit."
          },
          "415": {
            "description": "Unsupported, corrupt, disguised, or unsafe image."
          },
          "429": {
            "description": "Per-user upload rate limit exceeded."
          },
          "503": {
            "description": "Private storage is unavailable."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Create a transaction from a strictly validated private receipt.",
        "tags": [
          "transactions-v1"
        ]
      }
    },
...
    "/api/v1/transactions/{transaction_id}/analyses": {
      "parameters": [
        {
          "in": "path",
          "name": "transaction_id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ],
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisStartRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStartEnvelope"
                }
              }
            },
            "description": "Accepted"
          },
          "409": {
            "description": "OCR review is required before analysis."
          },
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "503": {
            "description": "Stored verification configuration is unavailable."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Run combined analysis or an OCR-result-only screenshot analysis.",
        "tags": [
          "ocr-v1"
        ]
      }
    },
...
    "/api/v1/transactions/{transaction_id}/ocr": {
      "parameters": [
        {
          "in": "path",
          "name": "transaction_id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ],
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRReviewEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Missing or invalid idempotency key."
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The USER role is required."
          },
          "404": {
            "description": "Transaction not found."
          },
          "409": {
            "description": "Invalid state, idempotency conflict or hash mismatch."
          },
          "429": {
            "description": "Per-user OCR rate limit exceeded."
          },
          "503": {
            "description": "Private storage or OCR processing is unavailable."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Run a bounded, reconstructable OCR pipeline for an owned receipt.",
        "tags": [
          "ocr-v1"
        ]
      }
    },
    "/api/v1/transactions/{transaction_id}/ocr-confirmations": {
      "parameters": [
        {
          "in": "path",
          "name": "transaction_id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ],
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OCRConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRConfirmationEnvelope"
                }
              }
            },
            "description": "A completed idempotent confirmation was replayed."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRConfirmationEnvelope"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "description": "Missing or invalid idempotency key."
          },
          "404": {
            "description": "Transaction or OCR result not found."
          },
          "409": {
            "description": "Invalid review state or idempotency conflict."
          },
          "422": {
            "description": "Canonical field or correction-reason validation failed."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Create an immutable canonical snapshot and correction audit.",
        "tags": [
          "ocr-v1"
        ]
      }
    },
    "/api/v1/transactions/{transaction_id}/ocr-review": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OCRReviewEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The USER role is required."
          },
          "404": {
            "description": "Transaction not found."
          },
          "409": {
            "description": "OCR has not been run."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Return the latest owner-safe OCR projection without raw token boxes.",
        "tags": [
          "ocr-v1"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "transaction_id",
          "required": true,
          "schema": {
            "format": "uuid",
            "type": "string"
          }
        }
      ]
    },
...
  },
...
}